api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
tortoise-orm==0.20.0
asyncpg==0.29.0